+++ hydra-cardano-api/src/Hydra/Cardano/Api.hs
@@ -124,7 +124,7 @@ import Hydra.Cardano.Api.ExecutionUnits as Extras
 import Hydra.Cardano.Api.Hash as Extras
 import Hydra.Cardano.Api.KeyWitness as Extras
 import Hydra.Cardano.Api.Lovelace as Extras
-import Hydra.Cardano.Api.MultiAssetSupportedInEra as Extras
+import Hydra.Cardano.Api.MaryEraOnwards as Extras
 import Hydra.Cardano.Api.NetworkId ()
 import Hydra.Cardano.Api.PlutusScript as Extras
 import Hydra.Cardano.Api.PolicyId as Extras